framework/media: update the configuration of audio gain/volume
- CONFIG_AUDIO_EXCLUDE_VOLUME is seperated into two configs for gain and volume in alc5658.c.
- AUDIO_MANAGER_VOLUME_NOT_SUPPORT is returned if CONFIG_AUDIO_EXCLUDE_VOLUME is set.
- 'defconfig' of artik055s/audio configuration is updated.
- Relevant parts in audio_manager.c and alc5658.c are updated.

+++ framework/include/media/MediaRecorder.h
@@ -156,7 +156,7 @@ public:
 	 * @return The result of setting the mic volume
 	 * @since TizenRT v2.0 PRE
 	 */
-	recorder_result_t setVolume(int vol);
+	recorder_result_t setVolume(uint8_t vol);
 	
 	/**
 	 * @brief Sets the DatSource of output data
